requirements.txt
streamlit
ultralytics
PyPDF2
pandas
Pillow
pymupdf